requirements.txt
gradio
ollama
PyYAML
chromadb
langchain